requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
pydantic>=2.5.0
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.6
bcrypt>=4.0.1
alembic>=1.12.0
pytest>=7.4.2
httpx>=0.24.1
python-dotenv>=1.0.0
email-validator>=2.0.0
